requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
asgiref==3.6.0
    # via
    #   -r requirements/base.txt
    #   django
astroid==2.12.13
    # via
    #   pylint
    #   pylint-celery
attrs==22.2.0
    # via pytest
certifi==2022.12.7
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.15.1
    # via
    #   -r requirements/base.txt
    #   cryptography
    #   pynacl
charset-normalizer==2.1.1
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.3
    # via
    #   -r requirements/base.txt
    #   click-log
    #   code-annotations
    #   edx-django-utils
    #   edx-lint
click-log==0.4.0
    # via edx-lint
code-annotations==1.3.0
    # via
    #   -r requirements/test.in
    #   edx-lint
coreapi==2.3.3
    # via
    #   -r requirements/base.txt
    #   django-rest-swagger
    #   openapi-codec
coreschema==0.0.4
    # via
    #   -r requirements/base.txt
    #   coreapi
coverage[toml]==7.0.3
    # via
    #   -r requirements/test.in
    #   pytest-cov
cryptography==39.0.0
    # via
    #   -r requirements/base.txt
    #   pyjwt
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   -r requirements/base.txt
    #   python3-openid
    #   social-auth-core
dill==0.3.6
    # via pylint
distlib==0.3.6
    # via virtualenv
dj-stripe==2.7.3
    # via -r requirements/base.txt
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   dj-stripe
    #   django-cors-headers
    #   django-crum
    #   django-extensions
    #   django-model-utils
    #   djangorestframework
    #   drf-jwt
    #   edx-auth-backends
    #   edx-django-release-util
    #   edx-django-utils
    #   edx-drf-extensions
    #   jsonfield
django-cors-headers==3.13.0
    # via -r requirements/base.txt
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-dynamic-fixture==3.1.2
    # via -r requirements/test.in
django-extensions==3.2.1
    # via -r requirements/base.txt
django-model-utils==4.3.1
    # via -r requirements/base.txt
django-rest-swagger==2.2.0
    # via -r requirements/base.txt
django-waffle==3.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-drf-extensions
djangorestframework==3.14.0
    # via
    #   -r requirements/base.txt
    #   django-rest-swagger
    #   drf-jwt
    #   edx-drf-extensions
drf-jwt==1.19.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-auth-backends==4.1.0
    # via -r requirements/base.txt
edx-django-release-util==1.2.0
    # via -r requirements/base.txt
edx-django-utils==5.2.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   edx-rest-api-client
edx-drf-extensions==8.4.0
    # via -r requirements/base.txt
edx-lint==5.3.0
    # via -r requirements/test.in
edx-opaque-keys==2.3.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-rest-api-client==5.5.0
    # via -r requirements/base.txt
exceptiongroup==1.1.0
    # via pytest
filelock==3.9.0
    # via
    #   tox
    #   virtualenv
future==0.18.2
    # via
    #   -r requirements/base.txt
    #   pyjwkest
idna==3.4
    # via
    #   -r requirements/base.txt
    #   requests
iniconfig==1.1.1
    # via pytest
isort==5.11.4
    # via pylint
itypes==1.2.0
    # via
    #   -r requirements/base.txt
    #   coreapi
jinja2==3.1.2
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   coreschema
jsonfield==3.1.0
    # via
    #   -r requirements/base.txt
    #   dj-stripe
lazy-object-proxy==1.9.0
    # via astroid
markupsafe==2.1.1
    # via
    #   -r requirements/base.txt
    #   jinja2
mccabe==0.7.0
    # via pylint
mock==5.0.0
    # via -r requirements/test.in
mysqlclient==2.1.1
    # via -r requirements/base.txt
newrelic==8.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
oauthlib==3.2.2
    # via
    #   -r requirements/base.txt
    #   requests-oauthlib
    #   social-auth-core
openapi-codec==1.3.2
    # via
    #   -r requirements/base.txt
    #   django-rest-swagger
packaging==22.0
    # via
    #   pytest
    #   tox
pbr==5.11.0
    # via
    #   -r requirements/base.txt
    #   stevedore
platformdirs==2.6.2
    # via
    #   pylint
    #   virtualenv
pluggy==1.0.0
    # via
    #   pytest
    #   tox
psutil==5.9.4
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
py==1.11.0
    # via tox
pycparser==2.21
    # via
    #   -r requirements/base.txt
    #   cffi
pycryptodomex==3.16.0
    # via
    #   -r requirements/base.txt
    #   pyjwkest
pyjwkest==1.4.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
pyjwt[crypto]==2.6.0
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   edx-auth-backends
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   social-auth-core
pylint==2.15.9
    # via
    #   edx-lint
    #   pylint-celery
    #   pylint-django
    #   pylint-plugin-utils
pylint-celery==0.3
    # via edx-lint
pylint-django==2.5.3
    # via edx-lint
pylint-plugin-utils==0.7
    # via
    #   pylint-celery
    #   pylint-django
pymongo==3.13.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pytest==7.2.0
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==4.0.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
python-dateutil==2.8.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
python-slugify==7.0.0
    # via code-annotations
python3-openid==3.2.0
    # via
    #   -r requirements/base.txt
    #   social-auth-core
pytz==2022.7
    # via
    #   -r requirements/base.txt
    #   django
    #   djangorestframework
pyyaml==6.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-release-util
requests==2.28.1
    # via
    #   -r requirements/base.txt
    #   coreapi
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   pyjwkest
    #   requests-oauthlib
    #   slumber
    #   social-auth-core
    #   stripe
requests-oauthlib==1.3.1
    # via
    #   -r requirements/base.txt
    #   social-auth-core
semantic-version==2.10.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
simplejson==3.18.1
    # via
    #   -r requirements/base.txt
    #   django-rest-swagger
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   django-dynamic-fixture
    #   edx-auth-backends
    #   edx-django-release-util
    #   edx-drf-extensions
    #   edx-lint
    #   pyjwkest
    #   python-dateutil
    #   tox
slumber==0.7.1
    # via
    #   -r requirements/base.txt
    #   edx-rest-api-client
social-auth-app-django==5.0.0
    # via
    #   -r requirements/base.txt
    #   edx-auth-backends
social-auth-core==4.3.0
    # via
    #   -r requirements/base.txt
    #   edx-auth-backends
    #   social-auth-app-django
sqlparse==0.4.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==4.1.1
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
stripe==4.2.0
    # via
    #   -r requirements/base.txt
    #   dj-stripe
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pylint
    #   pytest
    #   tox
tomlkit==0.11.6
    # via pylint
tox==3.28.0
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/test.in
typing-extensions==4.4.0
    # via
    #   astroid
    #   pylint
uritemplate==4.1.1
    # via
    #   -r requirements/base.txt
    #   coreapi
urllib3==1.26.13
    # via
    #   -r requirements/base.txt
    #   requests
virtualenv==20.17.1
    # via tox
wrapt==1.14.1
    # via astroid
